Extract traffic flow diagram logic
